backend/requirements.txt
# AWS Lambda dependencies
boto3>=1.26.0
PyMuPDF>=1.24.0
Pillow>=10.0.0

# Donut model dependencies
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

opencv-python-headless==4.8.1.78
pytesseract==0.3.10
PyPDF2==3.0.1
numpy>=1.23.5,<2.0
